LayoutTests/fast/events/wheel/continuous-platform-wheelevent-in-scrolling-div-expected.txt

PASS event.wheelDeltaY is window.expectedScrollTop*-3
PASS event.wheelDeltaX is window.expectedScrollLeft*-3
PASS event.wheelDelta is window.expectedScrollTop*-3
PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
